clean up some item tags
