+++ src/main/resources/data/dragonsurvival/recipe/charged_coal_1.json
@@ -7,7 +7,7 @@
   ],
   "key": {
     "C": {
-      "item": "minecraft:redstone"
+      "tag": "c:dusts/redstone"
     },
     "S": {
       "tag": "minecraft:coals"
