+++ src/main/resources/data/dragonsurvival/recipe/diamond_chorus.json
@@ -23,7 +23,7 @@
             "item": "minecraft:chorus_fruit"
         },
         {
-            "item": "minecraft:diamond"
+            "tag": "c:gems/diamond"
         }
     ],
     "result": {
