+++ src/main/resources/data/dragonsurvival/recipe/double_quartz.json
@@ -2,7 +2,7 @@
     "type": "minecraft:crafting_shapeless",
     "ingredients": [
         {
-            "item": "minecraft:quartz"
+            "tag": "c:gems/quartz"
         },
         {
             "item": "minecraft:amethyst_shard"
