+++ src/main/resources/data/dragonsurvival/recipe/explosive_copper.json
@@ -2,10 +2,10 @@
     "type": "minecraft:crafting_shapeless",
     "ingredients": [
         {
-            "item": "minecraft:copper_ingot"
+            "tag": "c:ingots/copper"
         },
         {
-            "item": "minecraft:gunpowder"
+            "tag": "c:gunpowders"
         }
     ],
     "result": {
