+++ src/main/resources/data/dragonsurvival/recipe/treasure_copper.json
