--- src/main/resources/data/dragonsurvival/recipe/treasure_copper_1.json
@@ -1,15 +1,0 @@
-{
-  "type": "crafting_shapeless",
-  "ingredients": [
-    {
-      "tag": "dragonsurvival:copper"
-    },
-    {
-      "item": "dragonsurvival:elder_dragon_dust"
-    }
-  ],
-  "result": {
-    "id": "dragonsurvival:treasure_copper",
-    "count": 1
-  }
-}
