+++ src/main/resources/data/dragonsurvival/recipe/treasure_diamond.json
@@ -2,7 +2,7 @@
   "type": "crafting_shapeless",
   "ingredients": [
     {
-      "item": "minecraft:diamond"
+      "tag": "c:gems/diamond"
     },
     {
       "item": "dragonsurvival:elder_dragon_dust"
