+++ src/main/resources/data/dragonsurvival/recipe/treasure_emerald.json
@@ -2,7 +2,7 @@
   "type": "crafting_shapeless",
   "ingredients": [
     {
-      "item": "minecraft:emerald"
+      "tag": "c:gems/emerald"
     },
     {
       "item": "dragonsurvival:elder_dragon_dust"
