+++ src/main/resources/data/dragonsurvival/recipe/treasure_iron.json
@@ -2,7 +2,7 @@
   "type": "crafting_shapeless",
   "ingredients": [
     {
-      "item": "minecraft:iron_nugget"
+      "tag": "c:nuggets/iron"
     },
     {
       "item": "dragonsurvival:elder_dragon_dust"
