+++ src/main/resources/data/dragonsurvival/tags/item/cooked_meats.json
@@ -1,11 +1,6 @@
 {
-  "replace": false,
   "values": [
-    "minecraft:cooked_beef",
-    "minecraft:cooked_chicken",
-    "minecraft:cooked_mutton",
-    "minecraft:cooked_porkchop",
-    "minecraft:cooked_rabbit",
+    "#c:foods/cooked_meats",
     "minecraft:rabbit_stew",
     {
       "id": "koopascritters:cooked_wild_bird",
