+++ src/main/resources/data/dragonsurvival/tags/item/cooked_seafood.json
@@ -1,8 +1,6 @@
 {
-  "replace": false,
   "values": [
-    "minecraft:cooked_cod",
-    "minecraft:cooked_salmon",
+    "#c:foods/cooked_fishes",
     {
       "id": "abnormals_delight:cooked_perch_slice",
       "required": false
